Get FilePath

+++ app/src/main/java/model/ForeignData.java
@@ -8,13 +8,13 @@ package model;
 public class ForeignData implements java.io.Serializable{
     //DateiMemo dateiMemo;
     private static final long serialVersionUID = -3180520592609573364L;
-    private long uid;
+    private int uid;
     private int fotoId;
     private double punktX;
     private double punktY;
     private String foreignIp;
 
-    public ForeignData( String foreignIp, double punktX, double punktY, int fotoId, long uid){
+    public ForeignData( String foreignIp, double punktX, double punktY, int fotoId, int uid){
         this.uid = uid;
         this.fotoId = fotoId;
         this.punktX = punktX;
@@ -28,11 +28,11 @@ public class ForeignData implements java.io.Serializable{
     }
 
 
-    public long getUid() {
+    public int getUid() {
         return uid;
     }
 
-    public void setUid(long uid) {
+    public void setUid(int uid) {
         this.uid = uid;
     }
